Correct toasts appear with static user values
So toasts appear when:
-Fields are empty on registration
-Fields do not match static fields
-Fields correctly match static fields

What isn't working is the making of dynamic userValues. My understanding of scope and how editText sets scope is what's wrong. If this were C I would use a pointer to the val userValues in a function so editText could use the textWatchers to assign a val to the tempValues